separate UI layout from workflow code
