+++ index.html
@@ -101,8 +101,8 @@
             </div>
         </div>
         <div id="footer-bar">
-            <button class="mode-button btn-mode-editor active">Editor</button>
-            <button class="mode-button btn-mode-serial">Serial</button>
+            <button id="btn-mode-editor" class="mode-button active">Editor</button>
+            <button id="btn-mode-serial" class="mode-button">Serial</button>
             <div class="spacer"></div>
             <button class="purple-button btn-info" disabled>Info<i class="fa-solid fa-info-circle"></i></button>
         </div>
