+++ js/state.js
@@ -1,0 +1,4 @@
+export default {
+    terminal: null,
+    fitter: null,
+}
